poke: Fix the reproducibility of installed scripts
There is a race condition between 2 actions at install around the
installed script in $pkgdatadir:
* Removal of existing scripts /usr/share/poke/*.pk
* Installation of default scripts in the same directory

Sadly, those 2 actions are not ordered. Depending on the build system load,
removal can (rarely) happen *after* the installation. In this case, no
script in present in /usr/share/poke/ when the install process end.

In OE, we always install on an empty sysroot, so, no pre-existing scripts can
exists. Remove the code that does the cleanup to avoid the
race-condition and ensure the scripts are always present.

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>
Signed-off-by: Khem Raj <raj.khem@gmail.com>